feat: 이미지 패치 로직을 useFetchImages 훅으로 분리
